povary/apps/ingredients/autocomplete_light_registry.py
import autocomplete_light
from ingredients.models import USAIngredient


class USAIngredientsAutocomplete(autocomplete_light.AutocompleteModelBase):
    search_fields = ['name_rus', ]

autocomplete_light.register(USAIngredient, USAIngredientsAutocomplete)